@@ -17,8 +17,8 @@
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/adapter-cloudflare": "^4.7.0",
 		"@sveltejs/adapter-static": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/jssha": "^3.0.0",
 		"@types/phoenix": "^1.6.0",
 		"@typescript-eslint/eslint-plugin": "^5.60.0",
@@ -27,19 +27,19 @@
 		"autoprefixer": "^10.4.14",
 		"eslint": "^8.43.0",
 		"eslint-config-prettier": "^8.8.0",
-		"eslint-plugin-svelte": "^2.31.1",
+		"eslint-plugin-svelte": "^2.45.1",
 		"kolorist": "^1.8.0",
 		"postcss": "^8.4.24",
 		"prettier": "^3.3.1",
 		"prettier-plugin-organize-imports": "^3.2.4",
-		"prettier-plugin-svelte": "^3.2.4",
+		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "^0.6.2",
-		"svelte": "^4.0.0",
-		"svelte-check": "^3.4.4",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"tailwindcss": "^3.3.2",
 		"tslib": "^2.5.3",
-		"typescript": "^5.1.3",
-		"vite": "^5.0.0",
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4",
 		"vite-plugin-url-copy": "^1.1.4",
 		"vitest": "^1.0.0",
 		"workbox-background-sync": "^7.0.0",
