@@ -1,8 +1,12 @@
 <script lang="ts">
 	import { nickname } from '$lib/stores/userinfo';
-	export let rank: number;
-	export let entryNickname: string;
-	export let stamps_collected: number;
+	interface Props {
+		rank: number;
+		entryNickname: string;
+		stamps_collected: number;
+	}
+
+	let { rank, entryNickname, stamps_collected }: Props = $props();
 
 	const ribbonIcon = `
 	<svg class="h-6 w-6" xmlns="http://www.w3.org/2000/svg" viewBox="0 -960 960 960" ><path d="m387-412 35-114-92-74h114l36-112 36 112h114l-93 74 35 114-92-71-93 71ZM240-40v-309q-38-42-59-96t-21-115q0-134 93-227t227-93q134 0 227 93t93 227q0 61-21 115t-59 96v309l-240-80-240 80Zm240-280q100 0 170-70t70-170q0-100-70-170t-170-70q-100 0-170 70t-70 170q0 100 70 170t170 70Z"/></svg>
