@@ -6,8 +6,8 @@
 	import { setToast } from '$lib/stores/toasts';
 	import { TOAST_TYPE } from '../../custom';
 	import { collectedStamps } from '$lib/stores/stamps';
-	let availableCameras = [] as ArrayLike<QrScanner.Camera>;
-	let sheetTorn = false;
+	let availableCameras = $state([] as ArrayLike<QrScanner.Camera>);
+	let sheetTorn = $state(false);
 
 	onMount(async () => {
 		availableCameras = await QrScanner.listCameras(true);
@@ -24,7 +24,7 @@
 	</ul>
 
 	<button
-		on:click={() => {
+		onclick={() => {
 			localStorage.clear();
 		}}
 		class="btn-error btn mx-auto w-8/12 gap-2 rounded-full"
@@ -38,19 +38,19 @@
 	<p class="text-xl font-bold">Stamps</p>
 	<p>Stamp sheet status: {sheetTorn ? 'torn' : 'untorn'}</p>
 	<button
-		on:click={() => localStorage.setItem('isStampSheetTorn', 'no')}
+		onclick={() => localStorage.setItem('isStampSheetTorn', 'no')}
 		class="btn-primary btn mx-auto w-8/12 gap-2 rounded-full text-base-100">Untear sheet</button
 	>
 	<button
 		class="btn-primary btn w-8/12 gap-2 rounded-full text-base-100"
-		on:click={async () => {
+		onclick={async () => {
 			await updateExpectedStamps();
 			setToast({ message: 'Stamp data refetched', type: TOAST_TYPE.SUCCESS });
 		}}>Refetch stamp data</button
 	>
 	<button
 		class="btn-error btn mx-auto w-8/12 gap-2 rounded-full"
-		on:click={() => {
+		onclick={() => {
 			collectedStamps.reset();
 			setToast({ message: 'Collected stamps reset', type: TOAST_TYPE.SUCCESS });
 		}}>Reset collected stamps</button
